[CX-2295]: refactoring
